perf: fused RMSNorm keeps the (1 + weight) add in fp32
The bf16 add rounded each channel scale by up to 0.4% and shifted the mean
logit by ~3% (probe gate caught it). Now identical to the reference to an
RMS relative error of 1.6e-5 with no bias.

(cherry picked from commit 485319450311e916a5caf713f08f14f7a723e661)
(cherry picked from commit 411a2e9b3b6e408a6bbea0e1913e70c34b60f892)

+++ src/transformers/models/qwen3_5/modeling_qwen3_5.py
@@ -865,8 +865,9 @@ class Qwen3_5RMSNorm(nn.Module):
 
     def forward(self, x):
         if _use_fla_norm(x):
-            # fused: fp32 rms-norm * (1 + weight), cast back to x.dtype; one Triton kernel each way
-            return _fla_rms_norm(x, 1.0 + self.weight, None, eps=self.eps)
+            # fused: fp32 rms-norm * (1 + weight), cast back to x.dtype; one Triton kernel each way.
+            # The add is done in fp32 like the reference: in bf16 it would round every channel scale by up to 0.4%.
+            return _fla_rms_norm(x, 1.0 + self.weight.float(), None, eps=self.eps)
         output = self._norm(x.float())
         # Llama does x.to(float16) * w whilst Qwen3_5 is (x * w).to(float16)
         # See https://github.com/huggingface/transformers/pull/29402
